// src/components/interfaces/IStock.ts
export interface IStock { 
    CHANGE_VALUE: number;
    PROFIT_LOSS_PERC: number; 
    PROFIT_LOSS: number; 
    MARKET_VALUE: number; 
    UNITS: number; 
    PRICE: number; 
    PURCHASE_PRICE: number;
    CATEGORY: string; 
    PREVIOUS_CLOSE: number;
    PRICE_MOVE: number; 
    SYMBOL: string;
    NAME: string;
    PRICE_MOVE_PERC: number; 
    DATETIME: string 
  };