// src/components/interfaces/IOverview.ts
export interface IOverview {
    TODAY_ICON: any;
    TOTAL_ICON: any;  
    KEY_TODAY: string;
    TODAY_LABEL: string;
    TODAY_PROFIT_LOSS: number;
    TODAY_COLOR: string;  
    TODAY_PROFIT_LOSS_PERC: number;  
    KEY_TOTAL: string;
    TOTAL_LABEL: string;
    TOTAL_PROFIT_LOSS: number;
    TOTAL_COLOR:string;
    TOTAL_PROFIT_LOSS_PERC: number;  
    KEY_HOLDINGS: string;
    HOLDINGS_COSTS: number;
    HOLDINGS_MARKET_VALUE: number;
    HOLDINGS_LABEL: string;
  }